pyproject.toml
[project]
name = "percolatte"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["pygame>=2.6.1", "pyglm>=2.8.3"]


[project.scripts]
main = "src.main:main"


[tool.pyright]
typeCheckingMode = "off"
